playback-test: Fix types for comparisons
Storing a 64 bit integer in a 32 bit integer and then checking
for the error cases might not be ideal.

error: comparison of constant -9223372036854775808 with
expression of type 'guint' (aka 'unsigned int') is always true

+++ tests/examples/playback/playback-test.c
@@ -2416,7 +2416,7 @@ buffer_size_activate_cb (GtkEntry * entry, PlaybackApp * app)
 
   text = gtk_entry_get_text (entry);
   if (text != NULL && *text != '\0') {
-    gint v;
+    gint64 v;
     gchar *endptr;
 
     v = g_ascii_strtoll (text, &endptr, 10);
@@ -2467,7 +2467,7 @@ connection_speed_activate_cb (GtkEntry * entry, PlaybackApp * app)
 
   text = gtk_entry_get_text (entry);
   if (text != NULL && *text != '\0') {
-    guint v;
+    gint64 v;
     gchar *endptr;
 
     v = g_ascii_strtoll (text, &endptr, 10);
